[arcane,accelerator] Move implementation for constituent commands on indexed selection in 'RunCommandMaterialEnumerator.h'.
